feat: add Works app
