+++ educations/views.py
@@ -1,11 +1,11 @@
 from django.urls import reverse_lazy
 from django.views.generic import CreateView, ListView, UpdateView, DeleteView
-from .models import Educations
-from .forms import EducationsForm
+from .models import Education
+from .forms import EducationForm
 
-class EducationsCreateView(CreateView):
-    model = Educations
-    form_class = EducationsForm
+class EducationCreateView(CreateView):
+    model = Education
+    form_class = EducationForm
     template_name = 'resume/my_education/create_education.html'
     success_url = reverse_lazy('resumes:edu-show')
 
@@ -15,18 +15,18 @@ class EducationsCreateView(CreateView):
         return super().form_valid(form)
 
 
-class EducationsListView(ListView):
-    model = Educations
+class EducationListView(ListView):
+    model = Education
     template_name = 'resume/my_education/show_education.html'
     context_object_name = 'educations' 
 
     def get_queryset(self):
-        return Educations.objects.filter(profile__user=self.request.user)
+        return Education.objects.filter(profile__user=self.request.user)
 
 
-class EducationsUpdateView(UpdateView):
-    model = Educations
-    form_class = EducationsForm
+class EducationUpdateView(UpdateView):
+    model = Education
+    form_class = EducationForm
     template_name = 'resume/my_education/update_education.html'
     success_url = reverse_lazy('resumes:edu-show')
     
@@ -34,7 +34,7 @@ class EducationsUpdateView(UpdateView):
         self.object = form.save()
         return super().form_valid(form)
 
-class EducationsDeleteView(DeleteView):
-    model = Educations
+class EducationDeleteView(DeleteView):
+    model = Education
     success_url = reverse_lazy('resumes:edu-show')
