+++ resume/urls.py
@@ -5,10 +5,6 @@ from .views import (
     ProfileCreateView,
     ProfileUpdateView,
     ProfileDeleteView,
-    WorkCreateView,
-    WorkListView,
-    WorkUpdateView,
-    WorkDeleteView,
     ProjectCreateView,
     ProjectListView,
     ProjectUpdateView,
@@ -16,10 +12,16 @@ from .views import (
     TotalListView,
 )
 from educations.views import (
-    EducationsCreateView,
-    EducationsListView,
-    EducationsUpdateView,
-    EducationsDeleteView,
+    EducationCreateView,
+    EducationListView,
+    EducationUpdateView,
+    EducationDeleteView,
+)
+from works.views import (
+    WorkCreateView,
+    WorkListView,
+    WorkUpdateView,
+    WorkDeleteView,
 )
 
 
@@ -30,18 +32,25 @@ urlpatterns = [
     path("create/", ProfileCreateView.as_view(), name="create"),
     path("edit/<pk>", ProfileUpdateView.as_view(), name="edit"),
     path("delete/<pk>", ProfileDeleteView.as_view(), name="delete"),
-    path("edu/", EducationsCreateView.as_view(), name="edu"),
-    path("edu/show/", EducationsListView.as_view(), name="edu-show"),
-    path("edu/edit/<pk>", EducationsUpdateView.as_view(), name="edu-edit"),
-    path("edu/delete/<pk>", EducationsDeleteView.as_view(), name="edu-delete"),
+    path("total/<int:profile_id>", TotalListView.as_view(), name="total"),
+
+
+    path("edu/", EducationCreateView.as_view(), name="edu"),
+    path("edu/show/", EducationListView.as_view(), name="edu-show"),
+    path("edu/edit/<pk>", EducationUpdateView.as_view(), name="edu-edit"),
+    path("edu/delete/<pk>", EducationDeleteView.as_view(), name="edu-delete"),
+
+
     path("work/", WorkCreateView.as_view(), name="work"),
     path("work/show/", WorkListView.as_view(), name="work-show"),
     path("work/edit/<pk>", WorkUpdateView.as_view(), name="work-edit"),
     path("work/delete/<pk>", WorkDeleteView.as_view(), name="work-delete"),
+
+    
     path("project/", ProjectCreateView.as_view(), name="project"),
     path("project/show/", ProjectListView.as_view(), name="project-show"),
     path("project/edit/<pk>", ProjectUpdateView.as_view(), name="project-edit"),
     path("project/delete/<pk>", ProjectDeleteView.as_view(), name="project-delete"),
-    path("total/<int:profile_id>", TotalListView.as_view(), name="total"),
+    
     path("<pk>/", ProfileListView.as_view(), name="show"),
 ]
