+++ resume/views.py
@@ -1,10 +1,12 @@
 from django.urls import reverse_lazy
 from django.views.generic import TemplateView,ListView,CreateView,UpdateView,DeleteView
-from .forms import ProfileForm, WorkForm,ProjectForm
-from .models import Profile, Work,Project
+from .forms import ProfileForm, ProjectForm
+from .models import Profile, Project
 from django.shortcuts import get_object_or_404
 from django.contrib.auth.mixins import LoginRequiredMixin
-from educations.models import Educations
+from educations.models import Education
+from works.models import Work
+
 
 
 class ResumeArea(TemplateView):
@@ -62,44 +64,6 @@ class ProfileDeleteView(DeleteView):
     success_url = reverse_lazy('resumes:index')
 
 
-
-
-
-class WorkCreateView(CreateView):
-    model = Work
-    form_class = WorkForm
-    template_name = 'resume/my_work/create_work.html'
-    success_url = reverse_lazy('resumes:work-show')
-
-    def form_valid(self, form):
-        self.object = form.save()
-        return super().form_valid(form)
-    
-
-class WorkListView(ListView):
-    model = Work
-    template_name = 'resume/my_work/show_work.html'
-    context_object_name = 'works' 
-
-    def get_queryset(self):
-        return Work.objects.filter(profile__user=self.request.user)
-    
-
-class WorkUpdateView(UpdateView):
-    model = Work
-    form_class = WorkForm
-    template_name = 'resume/my_work/update_work.html'
-    success_url = reverse_lazy('resumes:work-show')
-
-    def form_valid(self, form):
-        self.object = form.save()
-        return super().form_valid(form)
-
-class WorkDeleteView(DeleteView):
-    model = Work
-    success_url = reverse_lazy('resumes:work-show')
-
-
 class ProjectCreateView(CreateView):
     model = Project
     form_class = ProjectForm
@@ -143,12 +107,12 @@ class TotalListView(ListView):
         profile_id = self.kwargs['profile_id']
 
         profile_data = Profile.objects.filter(profile_id=profile_id)
-        educations_data = Educations.objects.filter(profile_id=profile_id)
+        education_data = Education.objects.filter(profile_id=profile_id)
         work_data = Work.objects.filter(profile_id=profile_id)
         project_data = Project.objects.filter(profile_id=profile_id)
 
         total_data = {
-            'educations_data': educations_data,
+            'education_data': education_data,
             'work_data': work_data,
             'project_data': project_data,
             'profile_data':profile_data,
